r14772@iad (orig r8835):  tfoote | 2009-01-05 11:43:40 -0800
 adding checking for badly nomalized quaternions on import and export from tf as per ticket:735, with debug outputs to rosconsole
 r14773@iad (orig r8836):  stuglaser | 2009-01-05 11:50:53 -0800
 Fixed xacro bug involving spaces in math
 
 r14774@iad (orig r8837):  stuglaser | 2009-01-05 11:51:20 -0800
 Added a few basic tests for xacro
 
 r14776@iad (orig r8839):  eitanme | 2009-01-05 12:57:54 -0800
 Changed raytracing model for stage at 2.5cm resolution to be 1cm
 r14777@iad (orig r8840):  jleibs | 2009-01-05 14:34:14 -0800
 Moving dcam toward working for a single (non-stereo) dcam.
 
 r14778@iad (orig r8841):  rob_wheeler | 2009-01-05 14:58:59 -0800
 Remove debug printf and fix compiler warning
 
 r14779@iad (orig r8842):  meeussen | 2009-01-05 14:59:55 -0800
 Test was too strict for the number of callbacks
 r14781@iad (orig r8843):  jdchen | 2009-01-05 15:12:22 -0800
 Added compiler flag for sse2
 
 r14782@iad (orig r8844):  hsujohnhsu | 2009-01-05 15:15:32 -0800
 add flags to make coarser meshes.
 
 r14783@iad (orig r8845):  kurtkonolige | 2009-01-05 15:17:41 -0800
 
 r14784@iad (orig r8846):  meeussen | 2009-01-05 15:24:12 -0800
 give node consistant name
 r14786@iad (orig r8847):  jdchen | 2009-01-05 15:28:44 -0800
 add compile switch -msse3 to the following
 
 r14791@iad (orig r8848):  hsujohnhsu | 2009-01-05 16:04:57 -0800
 
  * update python test script to test odom and nav separately. split tests between odom and nav.
  * update 2dnav-stack launch scripts fo for robot_pose_ekf
  * update locations due to map shift.
 
 
 r14793@iad (orig r8849):  kurtkonolige | 2009-01-05 16:09:16 -0800
 added stats collection to visualodometer
 r14794@iad (orig r8850):  hsujohnhsu | 2009-01-05 16:23:18 -0800
 update pr2.xml to use coarser mesh ~4,000 triangles vs ~30,000 triangles.
 
 r14799@iad (orig r8851):  pmihelich | 2009-01-05 16:33:01 -0800
 Fixed C-descriptor Python bindings.
 r14800@iad (orig r8852):  stuglaser | 2009-01-05 16:33:54 -0800
 Changed the elbow axis to be consistent with the upper arm pitch axis.
 
 r14803@iad (orig r8853):  pantofaru | 2009-01-05 16:40:30 -0800
 Now using the median color within a box instead of the mean color.
 r14804@iad (orig r8854):  vijaypradeep | 2009-01-05 16:41:50 -0800
 Added a downsampling parameter to the point_cloud_assembler
 
 r14805@iad (orig r8855):  tfoote | 2009-01-05 16:44:17 -0800
 at least test case is compiling
 r14808@iad (orig r8856):  vijaypradeep | 2009-01-05 17:24:58 -0800
 subsample_factor parameter now also subsamples across scans
 
 r14811@iad (orig r8857):  tfoote | 2009-01-05 17:39:00 -0800
 compiling and loading
 r14812@iad (orig r8858):  jamesbowman | 2009-01-05 17:59:04 -0800
 New test case.
 
 
 r14813@iad (orig r8859):  meeussen | 2009-01-05 18:03:49 -0800
 particle tracker running again; more detection code
 r14814@iad (orig r8860):  garybradski | 2009-01-05 18:10:16 -0800
 Adding capture.launch so that you can roslaunch stereo capture
 
 r14816@iad (orig r8861):  tpratkanis | 2009-01-05 18:32:12 -0800
 Make keepalive times parameters.
 r14818@iad (orig r8862):  tpratkanis | 2009-01-05 18:39:04 -0800
 Fix an issue where the costmap would not load any obstacles at start.
 r14819@iad (orig r8863):  hsujohnhsu | 2009-01-05 18:42:38 -0800
 update goal reached criteria settings to parameters.
 
 
 r14820@iad (orig r8864):  tpratkanis | 2009-01-05 18:42:45 -0800
 Add more error messages, and add hack solution to issue 799.
 r14821@iad (orig r8865):  hsujohnhsu | 2009-01-05 18:44:54 -0800
  * udpate test to check angle/translation separately.
  * update MoveBase goal tolerance variables.
  * comment out failing tests until fixed.
 
 r14822@iad (orig r8866):  hsujohnhsu | 2009-01-05 18:47:56 -0800
 remove _convex, specified in openrave parser for now.
 r14825@iad (orig r8867):  jleibs | 2009-01-05 19:11:06 -0800
 Fixing the build for Caroline.
 
 r14826@iad (orig r8868):  veedee | 2009-01-05 20:58:19 -0800
 do some leaves bookkeeping
 
 r14828@iad (orig r8869):  tfoote | 2009-01-05 21:24:00 -0800
 adding a secret topic for Radu std_msgs::Empty over /reset_time will clear the tf cache
 r14829@iad (orig r8870):  tfoote | 2009-01-05 21:32:20 -0800
 adding a secret topic for Radu std_msgs::Empty over /reset_time will clear the tf cache
 r14830@iad (orig r8871):  veedee | 2009-01-05 22:17:34 -0800
 added masks for different channels 
 
 r14831@iad (orig r8872):  veedee | 2009-01-05 22:50:30 -0800
 changed to const int
 
 r14832@iad (orig r8873):  veedee | 2009-01-05 22:50:44 -0800
 added distance as a channel for SSF
 
 r14833@iad (orig r8874):  veedee | 2009-01-05 23:08:25 -0800
 good catch by Tully. i'm tired -> going home... :)
 
 r14838@iad (orig r8875):  jleibs | 2009-01-06 01:05:46 -0800
 Overhaul of assorted stereo relating functions due to move of stereoinfo into disparity info and change from byte to uint8.
 
 r14839@iad (orig r8876):  jleibs | 2009-01-06 01:09:20 -0800
 Managed to miss this file.
 
 r14840@iad (orig r8877):  rdiankov | 2009-01-06 02:53:48 -0800
 tweaking of manipulation scripts and data gathering script for laser/camera calibration
 r14841@iad (orig r8878):  rdiankov | 2009-01-06 02:54:58 -0800
 updated prf with new cameras, added checkerboard detection launch file
 r14842@iad (orig r8879):  meeussen | 2009-01-06 08:55:16 -0800
 set path in same way for all tests
 r14843@iad (orig r8880):  hsujohnhsu | 2009-01-06 09:50:33 -0800
 higher resolution (25mm) wg maps
 r14846@iad (orig r8881):  tpratkanis | 2009-01-06 10:02:32 -0800
 Remove hack, breaking gazebo and fixing stage. Also remove error messages.
 r14848@iad (orig r8882):  jleibs | 2009-01-06 10:35:37 -0800
 Missed one other conversion from byte to uint8.
 
 r14849@iad (orig r8883):  michaeca | 2009-01-06 10:39:59 -0800
 added folder vision/place_recognition/temp_fast_kmeans
 
 
 
 r14850@iad (orig r8884):  michaeca | 2009-01-06 10:41:23 -0800
 added vision/place_recognition/temp_fast_kmeans
 
 
 
 r14851@iad (orig r8885):  tpratkanis | 2009-01-06 10:57:42 -0800
 Fix goal rotation problem.
 r14853@iad (orig r8886):  veedee | 2009-01-06 11:15:19 -0800
 added ANN as a dependency, changef cloud_kdtree to use ANN for now (more stable)
 
 r14854@iad (orig r8887):  veedee | 2009-01-06 11:16:17 -0800
 whoops, forgot to change to shared
 
 r14855@iad (orig r8888):  jamesbowman | 2009-01-06 11:19:45 -0800
 Save/load, test fail.
 
 
 r14856@iad (orig r8889):  hsujohnhsu | 2009-01-06 11:21:01 -0800
 use 25mm map for these tests.
 


git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/branches/josh@8891 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/calonder_descriptor/src/test/directed_test.cpp
@@ -52,8 +52,8 @@ int main(int argc, char** argv)
   cl.train(base_set, rng, /*make_patch,*/ 25, 10, 1000, num_corners); // Only 20 views?
   cl.write(tree_name);
 
-  //typedef float SigType;
-  typedef uint8_t SigType;
+  typedef float SigType;
+  //typedef uint8_t SigType;
   typedef Promote<SigType>::type DistanceType;
   BruteForceMatcher<SigType, CvPoint> matcher(cl.classes());
